lesson09 mailroom_oo, refact add mailrm ctrl class
