+++ students/mmancini/lesson09/canned_all_thankyou_letters.txt
@@ -1,0 +1,6 @@
+Dear Jane Smith Thank You for your generous donations totaling $75:
+Dear Tom Adams Thank You for your generous donations totaling $100:
+Dear Helen Smalls Thank You for your generous donations totaling $60:
+Dear Ming Chan Thank You for your generous donations totaling $150:
+Dear Mary Jones Thank You for your generous donations totaling $30:
+Dear Sam Adams Thank You for your generous donations totaling $5000:
